pyproject.toml
[tool.poetry]
name = "umi-recorder-recorder"
version = "0.1.0"
description = ""
authors = ["RockkLee <sam880103@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
streamlink = "^6.3.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
